+++ src/util/Vecteur.cpp
@@ -1,7 +1,7 @@
 #include "util/Vecteur.h"
 namespace util
 {
-    Vecteur::Vecteur(int x , int y) : x(x) , y(y)
+    Vecteur::Vecteur(float x , float y) : x(x) , y(y)
     {
     }
